requirements.txt
pydantic==2.6.4
fastapi[all]==0.110.0
pytest
passlib
python-jose
python-dotenv
alembic
psycopg2-binary
ulid-py
SQLAlchemy==2.0.22
inflect
pytest-ordering
qrcode
pytest-pretty
Pillow
pydantic-settings
requests
sendgrid
google-cloud-storage
rapidfuzz
pandas
openpyxl
xlrd
pyxlsb
pdfkit
xlsxwriter
textblob
logfire
logfire[fastapi]
twilio
websockets